pyproject.toml
[tool.poetry]
name = "research-report-generator"
version = "0.1.0"
description = "ArXiv Report summary generator"
authors = ["MAJOR-WAGH <67504964+Major-wagh@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
llama-index = "^0.12.8"
llama-index-indices-managed-llama-cloud = "^0.6.3"
llama-parse = "^0.5.19"
llama-cloud = "^0.1.7"
arxiv = "^2.1.3"
groq = "^0.13.1"
aiohappyeyeballs = "^2.4.4"
aiohttp = "^3.11.11"
aiosignal = "^1.3.2"
annotated-types = "^0.7.0"
anyio = "^4.7.0"
attrs = "^24.3.0"
beautifulsoup4 = "^4.12.3"
build = "^1.2.2.post1"
CacheControl = "^0.14.1"
certifi = "^2024.12.14"
charset-normalizer = "^3.4.1"
cleo = "^2.1.0"
click = "^8.1.8"
colorama = "^0.4.6"
crashtest = "^0.4.1"
dataclasses-json = "^0.6.7"
Deprecated = "^1.2.15"
dirtyjson = "^1.0.8"
distlib = "^0.3.9"
distro = "^1.9.0"
docstring_parser = "^0.16"
dulwich = "^0.21.7"
fastjsonschema = "^2.21.1"
feedparser = "^6.0.11"
filelock = "^3.16.1"
filetype = "^1.2.0"
frozenlist = "^1.5.0"
fsspec = "^2024.12.0"
greenlet = "^3.1.1"
h11 = "^0.14.0"
httpcore = "^1.0.7"
httpx = "^0.28.1"
huggingface-hub = "^0.27.0"
idna = "^3.10"
installer = "^0.7.0"
instructor = "^1.7.2"
Jinja2 = "^3.1.5"
jiter = "^0.8.2"
joblib = "^1.4.2"
keyring = "^24.3.1"
llama-index-agent-openai = "^0.4.1"
llama-index-cli = "^0.4.0"
llama-index-core = "^0.12.8"
llama-index-embeddings-openai = "^0.3.1"
llama-index-llms-groq = "^0.3.1"
llama-index-llms-openai = "^0.3.12"
llama-index-llms-openai-like = "^0.3.3"
llama-index-multi-modal-llms-openai = "^0.4.1"
llama-index-program-openai = "^0.3.1"
llama-index-question-gen-openai = "^0.3.0"
llama-index-readers-file = "^0.4.1"
llama-index-readers-llama-parse = "^0.4.0"
markdown-it-py = "^3.0.0"
MarkupSafe = "^3.0.2"
marshmallow = "^3.23.2"
mdurl = "^0.1.2"
more-itertools = "^10.5.0"
msgpack = "^1.1.0"
multidict = "^6.1.0"
mypy-extensions = "^1.0.0"
nest-asyncio = "^1.6.0"
networkx = "^3.4.2"
nltk = "^3.9.1"
numpy = "^2.2.1"
openai = "^1.58.1"
packaging = "^24.2"
pandas = "^2.2.3"
pexpect = "^4.9.0"
pillow = "^11.0.0"
pkginfo = "^1.12.0"
platformdirs = "^4.3.6"
poetry = "^1.8.5"
poetry-core = "^1.9.1"
poetry-plugin-export = "^1.8.0"
propcache = "^0.2.1"
ptyprocess = "^0.7.0"
pydantic = "^2.10.4"
pydantic_core = "^2.27.2"
Pygments = "^2.18.0"
pypdf = "^5.1.0"
pyproject_hooks = "^1.2.0"
python-dateutil = "^2.9.0.post0"
python-dotenv = "^1.0.1"
pytz = "^2024.2"
pywin32-ctypes = "^0.2.3"
PyYAML = "^6.0.2"
RapidFuzz = "^3.11.0"
regex = "^2024.11.6"
requests = "^2.32.3"
requests-toolbelt = "^1.0.0"
rich = "^13.9.4"
safetensors = "^0.5.0"
setuptools = "^75.6.0"
sgmllib3k = "^1.0.0"
shellingham = "^1.5.4"
six = "^1.17.0"
sniffio = "^1.3.1"
soupsieve = "^2.6"
SQLAlchemy = "^2.0.36"
striprtf = "^0.0.26"
tenacity = "^9.0.0"
tiktoken = "^0.8.0"
tokenizers = "^0.21.0"
tomlkit = "^0.13.2"
tqdm = "^4.67.1"
transformers = "^4.47.1"
trove-classifiers = "^2024.10.21.16"
typer = "^0.15.1"
typing-inspect = "^0.9.0"
typing_extensions = "^4.12.2"
tzdata = "^2024.2"
urllib3 = "^2.3.0"
virtualenv = "^20.28.0"
wrapt = "^1.17.0"
yarl = "^1.18.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
